feat: Migrate bump-version script to ES module format
